frontend started. testing ongoing for create product UI function

+++ backend/controllers/product.controller.js
@@ -54,11 +54,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
    const { id } = req.params;
 
+   if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({ success: false, message: "Invalid Product Id" })
+   }
+
    try {
       await Product.findByIdAndDelete(id);
       res.status(200).json({ success: true, message: "Product Deleted" })
    } catch (error) {
       console.log("Error in deleting proucts:", error.message);
-      res.status(404).json({ success: false, message: "Product Not Found" })
+      res.status(500).json({ success: false, message: "Server error" })
    }
 }
